完成歌单接口🙄

+++ services/playlist.go
@@ -15,11 +15,19 @@ func AddPlaylist(name string, status uint64, userId uint64) (model.Playlist, err
 	return playlist, model.DB.Create(&playlist).Error
 }
 
-func SearchPlaylist(name string, pageSize int, page int) ([]model.Playlist, int64, error) {
+func SearchPlaylist(name string, pageSize int, page int, user *model.User) ([]model.Playlist, int64, error) {
 	var playlists []model.Playlist
 	name = fmt.Sprintf("%%%s%%", name)
 	var count int64
-	db := model.DB.Model(&model.Playlist{}).Where("name Like ?", name).Or("description Like ?", name).Count(&count)
+	db := model.DB
+	if user == nil {
+		db = model.DB.Model(&model.Playlist{}).Where("name Like ?", name).Or("description Like ?", name).Count(&count)
+	} else if user.ID == 0 {
+		db = model.DB.Model(&model.Playlist{}).Where("( name Like ? OR description Like ? ) AND status = ?", name, name, 1).Count(&count)
+
+	} else {
+		db = model.DB.Debug().Model(&model.Playlist{}).Where("( name Like ? or description Like ? ) and userId = ?", name, name, user.ID).Count(&count)
+	}
 	if page < 1 {
 		page = 1
 	}
@@ -31,11 +39,15 @@ func SearchPlaylist(name string, pageSize int, page int) ([]model.Playlist, int6
 	return playlists, count, nil
 }
 
-func DelPlaylist(id uint64) error {
+func DelPlaylist(id uint64, user *model.User) error {
 	playlist := model.Playlist{}
 	db := model.DB.First(&playlist, id)
 	if db.Error != nil {
 		return errors.New("歌单不存在")
+	}
+
+	if user != nil && playlist.UserID != user.ID {
+		return errors.New("越权！")
 	}
 
 	return db.Delete(&playlist).Error
